+++ src/style/padding.rs
@@ -7,15 +7,15 @@
 /// For more info on Styling, see [`kaolin::style`]
 #[derive(Default, Clone, Copy)]
 pub struct Padding {
-    pub left: f32,
-    pub right: f32,
-    pub top: f32,
-    pub bottom: f32,
+    pub left: f64,
+    pub right: f64,
+    pub top: f64,
+    pub bottom: f64,
 }
 
 impl Padding {
     /// Creates a new `Padding` instance with the specified values.
-    pub fn new(left: f32, right: f32, top: f32, bottom: f32) -> Self {
+    pub fn new(left: f64, right: f64, top: f64, bottom: f64) -> Self {
         Padding {
             left,
             right,
@@ -25,7 +25,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` instance with the same value for all sides.
-    pub fn all(value: f32) -> Self {
+    pub fn all(value: f64) -> Self {
         Padding {
             left: value,
             right: value,
@@ -35,7 +35,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` instance with the same value for the horizontal sides (left and right).
-    pub fn horizontal(value: f32) -> Self {
+    pub fn horizontal(value: f64) -> Self {
         Padding {
             left: value,
             right: value,
@@ -45,7 +45,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` instance with the same value for the vertical sides (top and bottom).
-    pub fn vertical(value: f32) -> Self {
+    pub fn vertical(value: f64) -> Self {
         Padding {
             left: 0.0,
             right: 0.0,
@@ -55,7 +55,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` with the same value for the horizontal sides (left and right) and the vertical sides (top and bottom).
-    pub fn hor_ver(hor: f32, ver: f32) -> Self {
+    pub fn hor_ver(hor: f64, ver: f64) -> Self {
         Padding {
             left: hor,
             right: hor,
@@ -65,7 +65,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` with only the left value set, all others sides set to 0.
-    pub fn left(value: f32) -> Self {
+    pub fn left(value: f64) -> Self {
         Padding {
             left: value,
             right: 0.0,
@@ -75,7 +75,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` with only the right value set, all others sides set to 0.
-    pub fn right(value: f32) -> Self {
+    pub fn right(value: f64) -> Self {
         Padding {
             left: 0.0,
             right: value,
@@ -85,7 +85,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` with only the top value set, all others sides set to 0.
-    pub fn top(value: f32) -> Self {
+    pub fn top(value: f64) -> Self {
         Padding {
             left: 0.0,
             right: 0.0,
@@ -95,7 +95,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` with only the bottom value set, all others sides set to 0.
-    pub fn bottom(value: f32) -> Self {
+    pub fn bottom(value: f64) -> Self {
         Padding {
             left: 0.0,
             right: 0.0,
@@ -105,12 +105,12 @@ impl Padding {
     }
 
     /// Returns the total horizontal padding (left + right). Used internally to calculate layout positions.
-    pub fn x(&self) -> f32 {
+    pub fn x(&self) -> f64 {
         self.left + self.right
     }
 
     /// Returns the total vertical padding (top + bottom). Used internally to calculate layout positions.
-    pub fn y(&self) -> f32 {
+    pub fn y(&self) -> f64 {
         self.top + self.bottom
     }
 }
